refactor(policy): split doctor modules
