+++ extensions/policy/src/doctor/register.ts
@@ -8,7 +8,6 @@ import {
   type HealthCheckContext,
   type HealthFinding,
 } from "openclaw/plugin-sdk/health";
-import { normalizeProviderId } from "openclaw/plugin-sdk/provider-model-shared";
 import { normalizeAgentId } from "openclaw/plugin-sdk/routing";
 import { isRecord, uniqueStrings } from "openclaw/plugin-sdk/string-coerce-runtime";
 import {
@@ -33,548 +32,39 @@ const loadFsPromisesModule = async () => {
   return await fsPromisesModulePromise;
 };
 
-const CHECK_IDS = {
-  policyAttestationMismatch: "policy/attestation-hash-mismatch",
-  policyDeniedChannelProvider: "policy/channels-denied-provider",
-  policyHashMismatch: "policy/policy-hash-mismatch",
-  policyInvalidFile: "policy/policy-jsonc-invalid",
-  policyMissingFile: "policy/policy-jsonc-missing",
-  policyDeniedMcpServer: "policy/mcp-denied-server",
-  policyUnapprovedMcpServer: "policy/mcp-unapproved-server",
-  policyDeniedModelProvider: "policy/models-denied-provider",
-  policyUnapprovedModelProvider: "policy/models-unapproved-provider",
-  policyPrivateNetworkAccess: "policy/network-private-access-enabled",
-  policyIngressDmPolicyUnapproved: "policy/ingress-dm-policy-unapproved",
-  policyIngressDmScopeUnapproved: "policy/ingress-dm-scope-unapproved",
-  policyIngressOpenGroupsDenied: "policy/ingress-open-groups-denied",
-  policyIngressGroupMentionRequired: "policy/ingress-group-mention-required",
-  policyGatewayNonLoopbackBind: "policy/gateway-non-loopback-bind",
-  policyGatewayAuthDisabled: "policy/gateway-auth-disabled",
-  policyGatewayRateLimitMissing: "policy/gateway-rate-limit-missing",
-  policyGatewayControlUiInsecure: "policy/gateway-control-ui-insecure",
-  policyGatewayTailscaleFunnel: "policy/gateway-tailscale-funnel",
-  policyGatewayRemoteEnabled: "policy/gateway-remote-enabled",
-  policyGatewayHttpEndpointEnabled: "policy/gateway-http-endpoint-enabled",
-  policyGatewayHttpUrlFetchUnrestricted: "policy/gateway-http-url-fetch-unrestricted",
-  policyAgentsWorkspaceAccessDenied: "policy/agents-workspace-access-denied",
-  policyAgentsToolNotDenied: "policy/agents-tool-not-denied",
-  policyToolsElevatedEnabled: "policy/tools-elevated-enabled",
-  policyToolsAlsoAllowMissing: "policy/tools-also-allow-missing",
-  policyToolsAlsoAllowUnexpected: "policy/tools-also-allow-unexpected",
-  policyToolsExecAskUnapproved: "policy/tools-exec-ask-unapproved",
-  policyToolsExecHostUnapproved: "policy/tools-exec-host-unapproved",
-  policyToolsExecSecurityUnapproved: "policy/tools-exec-security-unapproved",
-  policyToolsFsWorkspaceOnlyRequired: "policy/tools-fs-workspace-only-required",
-  policyToolsProfileUnapproved: "policy/tools-profile-unapproved",
-  policyToolsRequiredDenyMissing: "policy/tools-required-deny-missing",
-  policySandboxModeUnapproved: "policy/sandbox-mode-unapproved",
-  policySandboxBackendUnapproved: "policy/sandbox-backend-unapproved",
-  policySandboxContainerPostureUnobservable: "policy/sandbox-container-posture-unobservable",
-  policySandboxContainerHostNetworkDenied: "policy/sandbox-container-host-network-denied",
-  policySandboxContainerNamespaceJoinDenied: "policy/sandbox-container-namespace-join-denied",
-  policySandboxContainerMountModeRequired: "policy/sandbox-container-mount-mode-required",
-  policySandboxContainerRuntimeSocketMount: "policy/sandbox-container-runtime-socket-mount",
-  policySandboxContainerUnconfinedProfile: "policy/sandbox-container-unconfined-profile",
-  policySandboxBrowserCdpSourceRangeMissing: "policy/sandbox-browser-cdp-source-range-missing",
-  policyDataHandlingRedactionDisabled: "policy/data-handling-redaction-disabled",
-  policyDataHandlingTelemetryContentCapture: "policy/data-handling-telemetry-content-capture",
-  policyDataHandlingSessionRetentionNotEnforced:
-    "policy/data-handling-session-retention-not-enforced",
-  policyDataHandlingSessionTranscriptMemory:
-    "policy/data-handling-session-transcript-memory-enabled",
-  policySecretsUnmanagedProvider: "policy/secrets-unmanaged-provider",
-  policySecretsDeniedProviderSource: "policy/secrets-denied-provider-source",
-  policySecretsInsecureProvider: "policy/secrets-insecure-provider",
-  policyAuthProfileInvalidMetadata: "policy/auth-profile-invalid-metadata",
-  policyAuthProfileUnapprovedMode: "policy/auth-profile-unapproved-mode",
-  policyExecApprovalsMissing: "policy/exec-approvals-missing",
-  policyExecApprovalsInvalid: "policy/exec-approvals-invalid",
-  policyExecApprovalsDefaultSecurityUnapproved: "policy/exec-approvals-default-security-unapproved",
-  policyExecApprovalsAgentSecurityUnapproved: "policy/exec-approvals-agent-security-unapproved",
-  policyExecApprovalsAutoAllowSkillsEnabled: "policy/exec-approvals-auto-allow-skills-enabled",
-  policyExecApprovalsAllowlistMissing: "policy/exec-approvals-allowlist-missing",
-  policyExecApprovalsAllowlistUnexpected: "policy/exec-approvals-allowlist-unexpected",
-  policyMissingToolOwner: "policy/tools-missing-owner",
-  policyMissingToolRisk: "policy/tools-missing-risk-level",
-  policyMissingToolSensitivity: "policy/tools-missing-sensitivity-token",
-  policyUnknownToolRisk: "policy/tools-unknown-risk-level",
-  policyUnknownToolSensitivity: "policy/tools-unknown-sensitivity-token",
-} as const;
-
-export const POLICY_CHECK_IDS = [
-  CHECK_IDS.policyMissingFile,
-  CHECK_IDS.policyInvalidFile,
-  CHECK_IDS.policyHashMismatch,
-  CHECK_IDS.policyAttestationMismatch,
-  CHECK_IDS.policyDeniedChannelProvider,
-  CHECK_IDS.policyDeniedMcpServer,
-  CHECK_IDS.policyUnapprovedMcpServer,
-  CHECK_IDS.policyDeniedModelProvider,
-  CHECK_IDS.policyUnapprovedModelProvider,
-  CHECK_IDS.policyPrivateNetworkAccess,
-  CHECK_IDS.policyIngressDmPolicyUnapproved,
-  CHECK_IDS.policyIngressDmScopeUnapproved,
-  CHECK_IDS.policyIngressOpenGroupsDenied,
-  CHECK_IDS.policyIngressGroupMentionRequired,
-  CHECK_IDS.policyGatewayNonLoopbackBind,
-  CHECK_IDS.policyGatewayAuthDisabled,
-  CHECK_IDS.policyGatewayRateLimitMissing,
-  CHECK_IDS.policyGatewayControlUiInsecure,
-  CHECK_IDS.policyGatewayTailscaleFunnel,
-  CHECK_IDS.policyGatewayRemoteEnabled,
-  CHECK_IDS.policyGatewayHttpEndpointEnabled,
-  CHECK_IDS.policyGatewayHttpUrlFetchUnrestricted,
-  CHECK_IDS.policyAgentsWorkspaceAccessDenied,
-  CHECK_IDS.policyAgentsToolNotDenied,
-  CHECK_IDS.policyToolsProfileUnapproved,
-  CHECK_IDS.policyToolsFsWorkspaceOnlyRequired,
-  CHECK_IDS.policyToolsExecSecurityUnapproved,
-  CHECK_IDS.policyToolsExecAskUnapproved,
-  CHECK_IDS.policyToolsExecHostUnapproved,
-  CHECK_IDS.policyToolsElevatedEnabled,
-  CHECK_IDS.policyToolsAlsoAllowMissing,
-  CHECK_IDS.policyToolsAlsoAllowUnexpected,
-  CHECK_IDS.policyToolsRequiredDenyMissing,
-  CHECK_IDS.policySandboxModeUnapproved,
-  CHECK_IDS.policySandboxBackendUnapproved,
-  CHECK_IDS.policySandboxContainerPostureUnobservable,
-  CHECK_IDS.policySandboxContainerHostNetworkDenied,
-  CHECK_IDS.policySandboxContainerNamespaceJoinDenied,
-  CHECK_IDS.policySandboxContainerMountModeRequired,
-  CHECK_IDS.policySandboxContainerRuntimeSocketMount,
-  CHECK_IDS.policySandboxContainerUnconfinedProfile,
-  CHECK_IDS.policySandboxBrowserCdpSourceRangeMissing,
-  CHECK_IDS.policyDataHandlingRedactionDisabled,
-  CHECK_IDS.policyDataHandlingTelemetryContentCapture,
-  CHECK_IDS.policyDataHandlingSessionRetentionNotEnforced,
-  CHECK_IDS.policyDataHandlingSessionTranscriptMemory,
-  CHECK_IDS.policySecretsUnmanagedProvider,
-  CHECK_IDS.policySecretsDeniedProviderSource,
-  CHECK_IDS.policySecretsInsecureProvider,
-  CHECK_IDS.policyAuthProfileInvalidMetadata,
-  CHECK_IDS.policyAuthProfileUnapprovedMode,
-  CHECK_IDS.policyExecApprovalsMissing,
-  CHECK_IDS.policyExecApprovalsInvalid,
-  CHECK_IDS.policyExecApprovalsDefaultSecurityUnapproved,
-  CHECK_IDS.policyExecApprovalsAgentSecurityUnapproved,
-  CHECK_IDS.policyExecApprovalsAutoAllowSkillsEnabled,
-  CHECK_IDS.policyExecApprovalsAllowlistMissing,
-  CHECK_IDS.policyExecApprovalsAllowlistUnexpected,
-  CHECK_IDS.policyMissingToolRisk,
-  CHECK_IDS.policyUnknownToolRisk,
-  CHECK_IDS.policyMissingToolSensitivity,
-  CHECK_IDS.policyMissingToolOwner,
-  CHECK_IDS.policyUnknownToolSensitivity,
-] as const;
-
-export type PolicyStrictnessKind =
-  | "allowlist-subset"
-  | "denylist-superset"
-  | "ordered-string"
-  | "requires-true"
-  | "requires-false"
-  | "exact-list";
-
-export type PolicyEmptyListSemantics = "disabled" | "meaningful";
-
-export type PolicyScopeSelectorKind = "agentIds" | "channelIds";
-
-export type PolicyRuleMetadata = {
-  readonly policyPath: readonly string[];
-  readonly strictness: PolicyStrictnessKind;
-  readonly valueType: "boolean" | "channel-provider-deny-rules" | "string" | "string-list";
-  readonly checkIds: readonly (typeof POLICY_CHECK_IDS)[number][];
-  readonly emptyList?: PolicyEmptyListSemantics;
-  readonly allowedValues?: readonly string[];
-  readonly caseSensitive?: boolean;
-  readonly normalizeValues?: "model-provider";
-  readonly orderedValues?: readonly string[];
-  readonly scopeSelectors?: readonly PolicyScopeSelectorKind[];
-};
-
-const SANDBOX_CONTAINER_POLICY_RULES = [
-  {
-    key: "denyHostNetwork",
-    label: "host network posture",
-    checkIds: [CHECK_IDS.policySandboxContainerHostNetworkDenied],
-  },
-  {
-    key: "denyContainerNamespaceJoin",
-    label: "container namespace posture",
-    checkIds: [CHECK_IDS.policySandboxContainerNamespaceJoinDenied],
-  },
-  {
-    key: "requireReadOnlyMounts",
-    label: "container mount mode posture",
-    checkIds: [CHECK_IDS.policySandboxContainerMountModeRequired],
-  },
-  {
-    key: "denyContainerRuntimeSocketMounts",
-    label: "container runtime socket mount posture",
-    checkIds: [CHECK_IDS.policySandboxContainerRuntimeSocketMount],
-  },
-  {
-    key: "denyUnconfinedProfiles",
-    label: "container security profile posture",
-    checkIds: [CHECK_IDS.policySandboxContainerUnconfinedProfile],
-  },
-] as const;
-
-const SANDBOX_POLICY_RULE_METADATA = [
-  {
-    policyPath: ["sandbox", "requireMode"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policySandboxModeUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["off", "non-main", "all"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["sandbox", "allowBackends"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policySandboxBackendUnapproved],
-    emptyList: "disabled",
-    scopeSelectors: ["agentIds"],
-  },
-  ...SANDBOX_CONTAINER_POLICY_RULES.map((rule) => ({
-    policyPath: ["sandbox", "containers", rule.key] as const,
-    strictness: "requires-true" as const,
-    valueType: "boolean" as const,
-    checkIds: rule.checkIds,
-    scopeSelectors: ["agentIds"] as const,
-  })),
-  {
-    policyPath: ["sandbox", "browser", "requireCdpSourceRange"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policySandboxBrowserCdpSourceRangeMissing],
-    scopeSelectors: ["agentIds"],
-  },
-] as const satisfies readonly PolicyRuleMetadata[];
-
-export const POLICY_RULE_METADATA = [
-  {
-    policyPath: ["channels", "denyRules"],
-    strictness: "denylist-superset",
-    valueType: "channel-provider-deny-rules",
-    checkIds: [CHECK_IDS.policyDeniedChannelProvider],
-    emptyList: "meaningful",
-    caseSensitive: true,
-  },
-  {
-    policyPath: ["mcp", "servers", "allow"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyUnapprovedMcpServer],
-    emptyList: "disabled",
-    caseSensitive: true,
-  },
-  {
-    policyPath: ["mcp", "servers", "deny"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyDeniedMcpServer],
-    caseSensitive: true,
-  },
-  {
-    policyPath: ["models", "providers", "allow"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyUnapprovedModelProvider],
-    emptyList: "disabled",
-    normalizeValues: "model-provider",
-  },
-  {
-    policyPath: ["models", "providers", "deny"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyDeniedModelProvider],
-    normalizeValues: "model-provider",
-  },
-  {
-    policyPath: ["network", "privateNetwork", "allow"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyPrivateNetworkAccess],
-  },
-  {
-    policyPath: ["ingress", "session", "requireDmScope"],
-    strictness: "ordered-string",
-    valueType: "string",
-    orderedValues: ["main", "per-peer", "per-channel-peer", "per-account-channel-peer"],
-    checkIds: [CHECK_IDS.policyIngressDmScopeUnapproved],
-  },
-  {
-    policyPath: ["gateway", "exposure", "allowNonLoopbackBind"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyGatewayNonLoopbackBind],
-  },
-  {
-    policyPath: ["gateway", "exposure", "allowTailscaleFunnel"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyGatewayTailscaleFunnel],
-  },
-  {
-    policyPath: ["gateway", "auth", "requireAuth"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyGatewayAuthDisabled],
-  },
-  {
-    policyPath: ["gateway", "auth", "requireExplicitRateLimit"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyGatewayRateLimitMissing],
-  },
-  {
-    policyPath: ["gateway", "controlUi", "allowInsecure"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyGatewayControlUiInsecure],
-  },
-  {
-    policyPath: ["gateway", "remote", "allow"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyGatewayRemoteEnabled],
-  },
-  {
-    policyPath: ["gateway", "http", "denyEndpoints"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyGatewayHttpEndpointEnabled],
-    allowedValues: ["chatCompletions", "responses"],
-    caseSensitive: true,
-  },
-  {
-    policyPath: ["gateway", "http", "requireUrlAllowlists"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyGatewayHttpUrlFetchUnrestricted],
-  },
-  {
-    policyPath: ["agents", "workspace", "allowedAccess"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyAgentsWorkspaceAccessDenied],
-    emptyList: "disabled",
-    allowedValues: ["none", "ro", "rw"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["agents", "workspace", "denyTools"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyAgentsToolNotDenied],
-    allowedValues: ["exec", "process", "write", "edit", "apply_patch"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "profiles", "allow"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyToolsProfileUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["minimal", "coding", "messaging", "full"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "fs", "requireWorkspaceOnly"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyToolsFsWorkspaceOnlyRequired],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "exec", "allowSecurity"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyToolsExecSecurityUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["deny", "allowlist", "full"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "exec", "requireAsk"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyToolsExecAskUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["off", "on-miss", "always"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "exec", "allowHosts"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyToolsExecHostUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["auto", "sandbox", "gateway", "node"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "elevated", "allow"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyToolsElevatedEnabled],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "alsoAllow", "expected"],
-    strictness: "exact-list",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyToolsAlsoAllowMissing, CHECK_IDS.policyToolsAlsoAllowUnexpected],
-    emptyList: "meaningful",
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "denyTools"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyToolsRequiredDenyMissing],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["tools", "requireMetadata"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [
-      CHECK_IDS.policyMissingToolRisk,
-      CHECK_IDS.policyMissingToolSensitivity,
-      CHECK_IDS.policyMissingToolOwner,
-    ],
-    allowedValues: ["risk", "sensitivity", "owner"],
-  },
-  ...SANDBOX_POLICY_RULE_METADATA,
-  {
-    policyPath: ["ingress", "channels", "allowDmPolicies"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyIngressDmPolicyUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["pairing", "allowlist", "open", "disabled"],
-    scopeSelectors: ["channelIds"],
-  },
-  {
-    policyPath: ["ingress", "channels", "denyOpenGroups"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyIngressOpenGroupsDenied],
-    scopeSelectors: ["channelIds"],
-  },
-  {
-    policyPath: ["ingress", "channels", "requireMentionInGroups"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyIngressGroupMentionRequired],
-    scopeSelectors: ["channelIds"],
-  },
-  {
-    policyPath: ["dataHandling", "sensitiveLogging", "requireRedaction"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyDataHandlingRedactionDisabled],
-  },
-  {
-    policyPath: ["dataHandling", "telemetry", "denyContentCapture"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyDataHandlingTelemetryContentCapture],
-  },
-  {
-    policyPath: ["dataHandling", "retention", "requireSessionMaintenance"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyDataHandlingSessionRetentionNotEnforced],
-  },
-  {
-    policyPath: ["dataHandling", "memory", "denySessionTranscriptIndexing"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyDataHandlingSessionTranscriptMemory],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["secrets", "requireManagedProviders"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policySecretsUnmanagedProvider],
-  },
-  {
-    policyPath: ["secrets", "denySources"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policySecretsDeniedProviderSource],
-  },
-  {
-    policyPath: ["secrets", "allowInsecureProviders"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policySecretsInsecureProvider],
-  },
-
-  {
-    policyPath: ["execApprovals", "requireFile"],
-    strictness: "requires-true",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyExecApprovalsMissing],
-  },
-  {
-    policyPath: ["execApprovals", "defaults", "allowSecurity"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyExecApprovalsDefaultSecurityUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["deny", "allowlist", "full"],
-  },
-  {
-    policyPath: ["execApprovals", "agents", "allowSecurity"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyExecApprovalsAgentSecurityUnapproved],
-    emptyList: "disabled",
-    allowedValues: ["deny", "allowlist", "full"],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["execApprovals", "agents", "allowAutoAllowSkills"],
-    strictness: "requires-false",
-    valueType: "boolean",
-    checkIds: [CHECK_IDS.policyExecApprovalsAutoAllowSkillsEnabled],
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["execApprovals", "agents", "allowlist", "expected"],
-    strictness: "exact-list",
-    valueType: "string-list",
-    checkIds: [
-      CHECK_IDS.policyExecApprovalsAllowlistMissing,
-      CHECK_IDS.policyExecApprovalsAllowlistUnexpected,
-    ],
-    emptyList: "meaningful",
-    caseSensitive: true,
-    scopeSelectors: ["agentIds"],
-  },
-  {
-    policyPath: ["auth", "profiles", "requireMetadata"],
-    strictness: "denylist-superset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyAuthProfileInvalidMetadata],
-    allowedValues: ["provider", "mode"],
-  },
-  {
-    policyPath: ["auth", "profiles", "allowModes"],
-    strictness: "allowlist-subset",
-    valueType: "string-list",
-    checkIds: [CHECK_IDS.policyAuthProfileUnapprovedMode],
-    emptyList: "disabled",
-    allowedValues: ["api_key", "aws-sdk", "oauth", "token"],
-  },
-] as const satisfies readonly PolicyRuleMetadata[];
+import { createPolicyDoctorChecks } from "./checks.js";
+import {
+  CHECK_IDS,
+  POLICY_CHECK_IDS,
+  POLICY_RULE_METADATA,
+  SANDBOX_CONTAINER_POLICY_RULES,
+  type PolicyRuleMetadata,
+  type PolicyScopeSelectorKind,
+} from "./metadata.js";
+import { gatewayExposureFindings } from "./scopes/gateway.js";
+import {
+  mcpServerFindings,
+  modelProviderFindings,
+  networkFindings,
+} from "./scopes/model-network.js";
+import { isPolicyValueAtLeastAsStrict } from "./strictness.js";
+import {
+  ocPathSegment,
+  readPolicyBoolean,
+  readPolicyStringArray,
+  readString,
+  readStringList,
+} from "./utils.js";
+export {
+  POLICY_CHECK_IDS,
+  POLICY_RULE_METADATA,
+  SANDBOX_CONTAINER_POLICY_RULES,
+  type PolicyEmptyListSemantics,
+  type PolicyRuleMetadata,
+  type PolicyScopeSelectorKind,
+  type PolicyStrictnessKind,
+} from "./metadata.js";
+export { isPolicyValueAtLeastAsStrict } from "./strictness.js";
 
 const POLICY_RULES: readonly PolicyRuleMetadata[] = POLICY_RULE_METADATA;
 
@@ -651,69 +141,16 @@ export function registerPolicyDoctorChecks(host?: PolicyDoctorRegistrationHost):
     return;
   }
   const registerHealthCheck = host?.registerHealthCheck ?? registerPluginHealthCheck;
-  registerHealthCheck(policyMissingFileCheck);
-  registerHealthCheck(policyInvalidFileCheck);
-  registerHealthCheck(policyHashMismatchCheck);
-  registerHealthCheck(policyAttestationMismatchCheck);
-  registerHealthCheck(policyChannelsDeniedProviderCheck);
-  registerHealthCheck(policyMcpDeniedServerCheck);
-  registerHealthCheck(policyMcpUnapprovedServerCheck);
-  registerHealthCheck(policyModelsDeniedProviderCheck);
-  registerHealthCheck(policyModelsUnapprovedProviderCheck);
-  registerHealthCheck(policyNetworkPrivateAccessCheck);
-  registerHealthCheck(policyIngressDmPolicyUnapprovedCheck);
-  registerHealthCheck(policyIngressDmScopeUnapprovedCheck);
-  registerHealthCheck(policyIngressOpenGroupsDeniedCheck);
-  registerHealthCheck(policyIngressGroupMentionRequiredCheck);
-  registerHealthCheck(policyGatewayNonLoopbackBindCheck);
-  registerHealthCheck(policyGatewayAuthDisabledCheck);
-  registerHealthCheck(policyGatewayRateLimitMissingCheck);
-  registerHealthCheck(policyGatewayControlUiInsecureCheck);
-  registerHealthCheck(policyGatewayTailscaleFunnelCheck);
-  registerHealthCheck(policyGatewayRemoteEnabledCheck);
-  registerHealthCheck(policyGatewayHttpEndpointEnabledCheck);
-  registerHealthCheck(policyGatewayHttpUrlFetchUnrestrictedCheck);
-  registerHealthCheck(policyAgentsWorkspaceAccessDeniedCheck);
-  registerHealthCheck(policyAgentsToolNotDeniedCheck);
-  registerHealthCheck(policyToolsProfileUnapprovedCheck);
-  registerHealthCheck(policyToolsFsWorkspaceOnlyRequiredCheck);
-  registerHealthCheck(policyToolsExecSecurityUnapprovedCheck);
-  registerHealthCheck(policyToolsExecAskUnapprovedCheck);
-  registerHealthCheck(policyToolsExecHostUnapprovedCheck);
-  registerHealthCheck(policyToolsElevatedEnabledCheck);
-  registerHealthCheck(policyToolsAlsoAllowMissingCheck);
-  registerHealthCheck(policyToolsAlsoAllowUnexpectedCheck);
-  registerHealthCheck(policyToolsRequiredDenyMissingCheck);
-  registerHealthCheck(policySandboxModeUnapprovedCheck);
-  registerHealthCheck(policySandboxBackendUnapprovedCheck);
-  registerHealthCheck(policySandboxContainerPostureUnobservableCheck);
-  registerHealthCheck(policySandboxContainerHostNetworkDeniedCheck);
-  registerHealthCheck(policySandboxContainerNamespaceJoinDeniedCheck);
-  registerHealthCheck(policySandboxContainerMountModeRequiredCheck);
-  registerHealthCheck(policySandboxContainerRuntimeSocketMountCheck);
-  registerHealthCheck(policySandboxContainerUnconfinedProfileCheck);
-  registerHealthCheck(policySandboxBrowserCdpSourceRangeMissingCheck);
-  registerHealthCheck(policyDataHandlingRedactionDisabledCheck);
-  registerHealthCheck(policyDataHandlingTelemetryContentCaptureCheck);
-  registerHealthCheck(policyDataHandlingSessionRetentionNotEnforcedCheck);
-  registerHealthCheck(policyDataHandlingSessionTranscriptMemoryCheck);
-  registerHealthCheck(policySecretsUnmanagedProviderCheck);
-  registerHealthCheck(policySecretsDeniedProviderSourceCheck);
-  registerHealthCheck(policySecretsInsecureProviderCheck);
-  registerHealthCheck(policyAuthProfileInvalidMetadataCheck);
-  registerHealthCheck(policyAuthProfileUnapprovedModeCheck);
-  registerHealthCheck(policyExecApprovalsMissingCheck);
-  registerHealthCheck(policyExecApprovalsInvalidCheck);
-  registerHealthCheck(policyExecApprovalsDefaultSecurityUnapprovedCheck);
-  registerHealthCheck(policyExecApprovalsAgentSecurityUnapprovedCheck);
-  registerHealthCheck(policyExecApprovalsAutoAllowSkillsEnabledCheck);
-  registerHealthCheck(policyExecApprovalsAllowlistMissingCheck);
-  registerHealthCheck(policyExecApprovalsAllowlistUnexpectedCheck);
-  registerHealthCheck(policyToolsMissingRiskCheck);
-  registerHealthCheck(policyToolsUnknownRiskCheck);
-  registerHealthCheck(policyToolsMissingSensitivityCheck);
-  registerHealthCheck(policyToolsMissingOwnerCheck);
-  registerHealthCheck(policyToolsUnknownSensitivityCheck);
+  for (const check of createPolicyDoctorChecks({
+    channelIdsFromFindings,
+    disableChannels,
+    evaluatePolicy,
+    findingsForCheck,
+    workspaceRepairsDisabledResult,
+    workspaceRepairsEnabled,
+  })) {
+    registerHealthCheck(check);
+  }
   registered = true;
 }
 
@@ -730,719 +167,6 @@ export function evaluatePolicy(ctx: HealthCheckContext): Promise<PolicyEvaluatio
   policyEvaluationCache.set(ctx, next);
   return next;
 }
-
-const policyMissingFileCheck: HealthCheck = {
-  id: CHECK_IDS.policyMissingFile,
-  kind: "plugin",
-  description: "The enabled Policy plugin has a policy file to verify.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyMissingFile);
-  },
-};
-
-const policyHashMismatchCheck: HealthCheck = {
-  id: CHECK_IDS.policyHashMismatch,
-  kind: "plugin",
-  description: "The policy file matches the configured expected hash.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyHashMismatch);
-  },
-};
-
-const policyAttestationMismatchCheck: HealthCheck = {
-  id: CHECK_IDS.policyAttestationMismatch,
-  kind: "plugin",
-  description: "The current policy check matches the accepted attestation.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyAttestationMismatch);
-  },
-};
-
-const policyInvalidFileCheck: HealthCheck = {
-  id: CHECK_IDS.policyInvalidFile,
-  kind: "plugin",
-  description: "The enabled policy file parses before policy checks run.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyInvalidFile);
-  },
-};
-
-const policyChannelsDeniedProviderCheck: HealthCheck = {
-  id: CHECK_IDS.policyDeniedChannelProvider,
-  kind: "plugin",
-  description: "Configured channels satisfy policy deny rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyDeniedChannelProvider);
-  },
-  async repair(ctx, findings) {
-    if (!workspaceRepairsEnabled(ctx)) {
-      return workspaceRepairsDisabledResult("channel config");
-    }
-    const channelIds = channelIdsFromFindings(findings);
-    if (channelIds.length === 0) {
-      return {
-        status: "skipped",
-        reason: "no channel findings matched a configurable channel",
-        changes: [],
-      };
-    }
-    const next = disableChannels(ctx.cfg, channelIds);
-    if (next.changed.length === 0) {
-      return {
-        status: "skipped",
-        reason: "matching channels were already disabled or missing",
-        changes: [],
-      };
-    }
-    return {
-      config: next.config,
-      changes: next.changed.map((id) => `Disabled channels.${id}.enabled for policy conformance.`),
-    };
-  },
-};
-
-const policyMcpDeniedServerCheck: HealthCheck = {
-  id: CHECK_IDS.policyDeniedMcpServer,
-  kind: "plugin",
-  description: "Configured MCP servers do not match policy deny rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyDeniedMcpServer);
-  },
-};
-
-const policyMcpUnapprovedServerCheck: HealthCheck = {
-  id: CHECK_IDS.policyUnapprovedMcpServer,
-  kind: "plugin",
-  description: "Configured MCP servers do not match policy allow rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyUnapprovedMcpServer);
-  },
-};
-
-const policyModelsDeniedProviderCheck: HealthCheck = {
-  id: CHECK_IDS.policyDeniedModelProvider,
-  kind: "plugin",
-  description: "Configured model providers do not match policy deny rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyDeniedModelProvider);
-  },
-};
-
-const policyModelsUnapprovedProviderCheck: HealthCheck = {
-  id: CHECK_IDS.policyUnapprovedModelProvider,
-  kind: "plugin",
-  description: "Configured model providers do not match policy allow rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyUnapprovedModelProvider);
-  },
-};
-
-const policyNetworkPrivateAccessCheck: HealthCheck = {
-  id: CHECK_IDS.policyPrivateNetworkAccess,
-  kind: "plugin",
-  description: "Network SSRF policy settings match private-network requirements.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyPrivateNetworkAccess);
-  },
-};
-
-const policyIngressDmPolicyUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyIngressDmPolicyUnapproved,
-  kind: "plugin",
-  description: "Channel direct-message access policy matches ingress requirements.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyIngressDmPolicyUnapproved);
-  },
-};
-
-const policyIngressDmScopeUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyIngressDmScopeUnapproved,
-  kind: "plugin",
-  description: "Direct-message sessions use the policy-required isolation scope.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyIngressDmScopeUnapproved);
-  },
-};
-
-const policyIngressOpenGroupsDeniedCheck: HealthCheck = {
-  id: CHECK_IDS.policyIngressOpenGroupsDenied,
-  kind: "plugin",
-  description: "Channel group access does not use open group policy when denied.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyIngressOpenGroupsDenied);
-  },
-};
-
-const policyIngressGroupMentionRequiredCheck: HealthCheck = {
-  id: CHECK_IDS.policyIngressGroupMentionRequired,
-  kind: "plugin",
-  description: "Channel group access keeps mention gates enabled when required.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyIngressGroupMentionRequired);
-  },
-};
-
-const policyGatewayNonLoopbackBindCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayNonLoopbackBind,
-  kind: "plugin",
-  description: "Gateway bind posture matches policy exposure requirements.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyGatewayNonLoopbackBind);
-  },
-};
-
-const policyGatewayAuthDisabledCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayAuthDisabled,
-  kind: "plugin",
-  description: "Gateway authentication remains enabled when required by policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyGatewayAuthDisabled);
-  },
-};
-
-const policyGatewayRateLimitMissingCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayRateLimitMissing,
-  kind: "plugin",
-  description: "Gateway authentication rate-limit posture is explicit when required by policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyGatewayRateLimitMissing);
-  },
-};
-
-const policyGatewayControlUiInsecureCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayControlUiInsecure,
-  kind: "plugin",
-  description: "Gateway Control UI insecure exposure toggles remain disabled by policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyGatewayControlUiInsecure);
-  },
-};
-
-const policyGatewayTailscaleFunnelCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayTailscaleFunnel,
-  kind: "plugin",
-  description: "Gateway Tailscale Funnel exposure matches policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyGatewayTailscaleFunnel);
-  },
-};
-
-const policyGatewayRemoteEnabledCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayRemoteEnabled,
-  kind: "plugin",
-  description: "Remote gateway mode matches policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyGatewayRemoteEnabled);
-  },
-};
-
-const policyGatewayHttpEndpointEnabledCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayHttpEndpointEnabled,
-  kind: "plugin",
-  description: "Gateway HTTP API endpoints match policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyGatewayHttpEndpointEnabled);
-  },
-};
-
-const policyGatewayHttpUrlFetchUnrestrictedCheck: HealthCheck = {
-  id: CHECK_IDS.policyGatewayHttpUrlFetchUnrestricted,
-  kind: "plugin",
-  description: "Gateway HTTP URL-fetch inputs have allowlists when required by policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyGatewayHttpUrlFetchUnrestricted,
-    );
-  },
-};
-
-const policyAgentsWorkspaceAccessDeniedCheck: HealthCheck = {
-  id: CHECK_IDS.policyAgentsWorkspaceAccessDenied,
-  kind: "plugin",
-  description: "Agent sandbox workspace access matches policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyAgentsWorkspaceAccessDenied);
-  },
-};
-
-const policyAgentsToolNotDeniedCheck: HealthCheck = {
-  id: CHECK_IDS.policyAgentsToolNotDenied,
-  kind: "plugin",
-  description: "Agent workspace mutation/runtime tools are denied when policy requires it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyAgentsToolNotDenied);
-  },
-};
-
-const policyToolsProfileUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsProfileUnapproved,
-  kind: "plugin",
-  description: "Configured tool profiles match policy allow rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsProfileUnapproved);
-  },
-};
-
-const policyToolsFsWorkspaceOnlyRequiredCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsFsWorkspaceOnlyRequired,
-  kind: "plugin",
-  description: "Filesystem tools use workspace-only posture when policy requires it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyToolsFsWorkspaceOnlyRequired,
-    );
-  },
-};
-
-const policyToolsExecSecurityUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsExecSecurityUnapproved,
-  kind: "plugin",
-  description: "Exec tool security mode matches policy allow rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsExecSecurityUnapproved);
-  },
-};
-
-const policyToolsExecAskUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsExecAskUnapproved,
-  kind: "plugin",
-  description: "Exec tool ask mode matches policy allow rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsExecAskUnapproved);
-  },
-};
-
-const policyToolsExecHostUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsExecHostUnapproved,
-  kind: "plugin",
-  description: "Exec tool host routing matches policy allow rules.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsExecHostUnapproved);
-  },
-};
-
-const policyToolsElevatedEnabledCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsElevatedEnabled,
-  kind: "plugin",
-  description: "Elevated tool mode remains disabled when policy requires it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsElevatedEnabled);
-  },
-};
-
-const policyToolsAlsoAllowMissingCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsAlsoAllowMissing,
-  kind: "plugin",
-  description: "Configured tools.alsoAllow entries include policy expected lists.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsAlsoAllowMissing);
-  },
-};
-
-const policyToolsAlsoAllowUnexpectedCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsAlsoAllowUnexpected,
-  kind: "plugin",
-  description: "Configured tools.alsoAllow entries match policy expected lists.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsAlsoAllowUnexpected);
-  },
-};
-
-const policyToolsRequiredDenyMissingCheck: HealthCheck = {
-  id: CHECK_IDS.policyToolsRequiredDenyMissing,
-  kind: "plugin",
-  description: "Configured tool deny lists include tools required by policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyToolsRequiredDenyMissing);
-  },
-};
-
-const policySandboxModeUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxModeUnapproved,
-  kind: "plugin",
-  description: "Sandbox mode config satisfies policy requirements.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policySandboxModeUnapproved);
-  },
-};
-
-const policySandboxBackendUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxBackendUnapproved,
-  kind: "plugin",
-  description: "Sandbox backend config satisfies policy requirements.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policySandboxBackendUnapproved);
-  },
-};
-
-const policySandboxContainerPostureUnobservableCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxContainerPostureUnobservable,
-  kind: "plugin",
-  description: "Sandbox container posture policy only targets observable container backends.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policySandboxContainerPostureUnobservable,
-    );
-  },
-};
-
-const policySandboxContainerHostNetworkDeniedCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxContainerHostNetworkDenied,
-  kind: "plugin",
-  description: "Sandbox container config avoids host network mode.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policySandboxContainerHostNetworkDenied,
-    );
-  },
-};
-
-const policySandboxContainerNamespaceJoinDeniedCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxContainerNamespaceJoinDenied,
-  kind: "plugin",
-  description: "Sandbox container config avoids joining another container network namespace.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policySandboxContainerNamespaceJoinDenied,
-    );
-  },
-};
-
-const policySandboxContainerMountModeRequiredCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxContainerMountModeRequired,
-  kind: "plugin",
-  description: "Sandbox container mounts are read-only when policy requires it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policySandboxContainerMountModeRequired,
-    );
-  },
-};
-
-const policySandboxContainerRuntimeSocketMountCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxContainerRuntimeSocketMount,
-  kind: "plugin",
-  description: "Sandbox container mounts avoid host container runtime sockets.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policySandboxContainerRuntimeSocketMount,
-    );
-  },
-};
-
-const policySandboxContainerUnconfinedProfileCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxContainerUnconfinedProfile,
-  kind: "plugin",
-  description: "Sandbox container profile config avoids unconfined profiles.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policySandboxContainerUnconfinedProfile,
-    );
-  },
-};
-
-const policySandboxBrowserCdpSourceRangeMissingCheck: HealthCheck = {
-  id: CHECK_IDS.policySandboxBrowserCdpSourceRangeMissing,
-  kind: "plugin",
-  description: "Sandbox browser CDP config includes a source range when policy requires it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policySandboxBrowserCdpSourceRangeMissing,
-    );
-  },
-};
-
-const policyDataHandlingRedactionDisabledCheck: HealthCheck = {
-  id: CHECK_IDS.policyDataHandlingRedactionDisabled,
-  kind: "plugin",
-  description: "Sensitive logging redaction remains enabled when policy requires it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyDataHandlingRedactionDisabled,
-    );
-  },
-};
-
-const policyDataHandlingTelemetryContentCaptureCheck: HealthCheck = {
-  id: CHECK_IDS.policyDataHandlingTelemetryContentCapture,
-  kind: "plugin",
-  description: "Telemetry content capture remains disabled when policy denies it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyDataHandlingTelemetryContentCapture,
-    );
-  },
-};
-
-const policyDataHandlingSessionRetentionNotEnforcedCheck: HealthCheck = {
-  id: CHECK_IDS.policyDataHandlingSessionRetentionNotEnforced,
-  kind: "plugin",
-  description: "Session retention maintenance is enforced when policy requires it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyDataHandlingSessionRetentionNotEnforced,
-    );
-  },
-};
-
-const policyDataHandlingSessionTranscriptMemoryCheck: HealthCheck = {
-  id: CHECK_IDS.policyDataHandlingSessionTranscriptMemory,
-  kind: "plugin",
-  description: "Session transcript memory indexing remains disabled when policy denies it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyDataHandlingSessionTranscriptMemory,
-    );
-  },
-};
-
-const policySecretsUnmanagedProviderCheck: HealthCheck = {
-  id: CHECK_IDS.policySecretsUnmanagedProvider,
-  kind: "plugin",
-  description:
-    "OpenClaw config SecretRefs use configured secret providers when policy requires managed providers.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policySecretsUnmanagedProvider);
-  },
-};
-
-const policySecretsDeniedProviderSourceCheck: HealthCheck = {
-  id: CHECK_IDS.policySecretsDeniedProviderSource,
-  kind: "plugin",
-  description:
-    "OpenClaw config secret providers and SecretRefs do not use sources denied by policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policySecretsDeniedProviderSource);
-  },
-};
-
-const policySecretsInsecureProviderCheck: HealthCheck = {
-  id: CHECK_IDS.policySecretsInsecureProvider,
-  kind: "plugin",
-  description:
-    "Configured secret providers do not opt into insecure posture unless policy allows it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policySecretsInsecureProvider);
-  },
-};
-
-const policyAuthProfileInvalidMetadataCheck: HealthCheck = {
-  id: CHECK_IDS.policyAuthProfileInvalidMetadata,
-  kind: "plugin",
-  description: "OpenClaw config auth profiles declare required provider and mode metadata.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyAuthProfileInvalidMetadata);
-  },
-};
-
-const policyAuthProfileUnapprovedModeCheck: HealthCheck = {
-  id: CHECK_IDS.policyAuthProfileUnapprovedMode,
-  kind: "plugin",
-  description: "OpenClaw config auth profile modes stay within the policy allowlist.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyAuthProfileUnapprovedMode);
-  },
-};
-
-const policyExecApprovalsMissingCheck: HealthCheck = {
-  id: CHECK_IDS.policyExecApprovalsMissing,
-  kind: "plugin",
-  description: "Required exec approvals artifact is present for policy conformance.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyExecApprovalsMissing);
-  },
-};
-
-const policyExecApprovalsInvalidCheck: HealthCheck = {
-  id: CHECK_IDS.policyExecApprovalsInvalid,
-  kind: "plugin",
-  description: "Exec approvals artifact parses before policy checks run.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyExecApprovalsInvalid);
-  },
-};
-
-const policyExecApprovalsDefaultSecurityUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyExecApprovalsDefaultSecurityUnapproved,
-  kind: "plugin",
-  description: "Exec approval defaults use a policy-approved security mode.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyExecApprovalsDefaultSecurityUnapproved,
-    );
-  },
-};
-
-const policyExecApprovalsAgentSecurityUnapprovedCheck: HealthCheck = {
-  id: CHECK_IDS.policyExecApprovalsAgentSecurityUnapproved,
-  kind: "plugin",
-  description: "Per-agent exec approval settings use policy-approved security modes.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyExecApprovalsAgentSecurityUnapproved,
-    );
-  },
-};
-
-const policyExecApprovalsAutoAllowSkillsEnabledCheck: HealthCheck = {
-  id: CHECK_IDS.policyExecApprovalsAutoAllowSkillsEnabled,
-  kind: "plugin",
-  description:
-    "Exec approval agents do not implicitly auto-allow skill CLIs unless policy allows it.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyExecApprovalsAutoAllowSkillsEnabled,
-    );
-  },
-};
-
-const policyExecApprovalsAllowlistMissingCheck: HealthCheck = {
-  id: CHECK_IDS.policyExecApprovalsAllowlistMissing,
-  kind: "plugin",
-  description: "Exec approval allowlists include every pattern required by policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyExecApprovalsAllowlistMissing,
-    );
-  },
-};
-
-const policyExecApprovalsAllowlistUnexpectedCheck: HealthCheck = {
-  id: CHECK_IDS.policyExecApprovalsAllowlistUnexpected,
-  kind: "plugin",
-  description: "Exec approval allowlists do not contain patterns outside policy.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(
-      await evaluatePolicy(ctx),
-      CHECK_IDS.policyExecApprovalsAllowlistUnexpected,
-    );
-  },
-};
-
-const policyToolsMissingRiskCheck: HealthCheck = {
-  id: CHECK_IDS.policyMissingToolRisk,
-  kind: "plugin",
-  description: "TOOLS.md policy entries declare explicit risk levels.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyMissingToolRisk);
-  },
-};
-
-const policyToolsUnknownRiskCheck: HealthCheck = {
-  id: CHECK_IDS.policyUnknownToolRisk,
-  kind: "plugin",
-  description: "TOOLS.md policy entries use known risk levels.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyUnknownToolRisk);
-  },
-};
-
-const policyToolsMissingSensitivityCheck: HealthCheck = {
-  id: CHECK_IDS.policyMissingToolSensitivity,
-  kind: "plugin",
-  description: "TOOLS.md policy entries declare default artifact sensitivity.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyMissingToolSensitivity);
-  },
-};
-
-const policyToolsUnknownSensitivityCheck: HealthCheck = {
-  id: CHECK_IDS.policyUnknownToolSensitivity,
-  kind: "plugin",
-  description: "TOOLS.md policy entries use known sensitivity levels.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyUnknownToolSensitivity);
-  },
-};
-
-const policyToolsMissingOwnerCheck: HealthCheck = {
-  id: CHECK_IDS.policyMissingToolOwner,
-  kind: "plugin",
-  description: "TOOLS.md policy entries declare an accountable owner.",
-  source: "policy",
-  async detect(ctx) {
-    return findingsForCheck(await evaluatePolicy(ctx), CHECK_IDS.policyMissingToolOwner);
-  },
-};
 
 async function evaluatePolicyUncached(ctx: HealthCheckContext): Promise<PolicyEvaluation> {
   const settings = policySettings(ctx);
@@ -3549,171 +2273,6 @@ function invalidChannelDenyRuleFindings(
   ];
 }
 
-function mcpServerFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  const denied = new Set(readStringList(policy, ["mcp", "servers", "deny"], { lowercase: false }));
-  const allowed = readStringList(policy, ["mcp", "servers", "allow"], { lowercase: false });
-  const allowedSet = new Set(allowed);
-  const findings: HealthFinding[] = [];
-
-  for (const server of evidence.mcpServers) {
-    if (denied.has(server.id)) {
-      findings.push({
-        checkId: CHECK_IDS.policyDeniedMcpServer,
-        severity: "error",
-        message: `MCP server '${server.id}' is denied by policy.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: server.source,
-        target: server.source,
-        requirement: `oc://${policyDocName}/mcp/servers/deny`,
-        fixHint: "Remove this configured MCP server or update the policy after review.",
-      });
-      continue;
-    }
-    if (allowedSet.size > 0 && !allowedSet.has(server.id)) {
-      findings.push({
-        checkId: CHECK_IDS.policyUnapprovedMcpServer,
-        severity: "error",
-        message: `MCP server '${server.id}' is not in the policy allowlist.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: server.source,
-        target: server.source,
-        requirement: `oc://${policyDocName}/mcp/servers/allow`,
-        fixHint: "Use an approved MCP server or update the policy after review.",
-      });
-    }
-  }
-
-  return findings;
-}
-
-function modelProviderFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  const denied = new Set(readModelProviderPolicyList(policy, ["models", "providers", "deny"]));
-  const allowed = readModelProviderPolicyList(policy, ["models", "providers", "allow"]);
-  const allowedSet = new Set(allowed);
-  const findings: HealthFinding[] = [];
-
-  for (const provider of evidence.modelProviders) {
-    findings.push(...modelProviderConformanceFindings(provider, denied, allowedSet, policyDocName));
-  }
-  for (const modelRef of evidence.modelRefs) {
-    findings.push(...modelRefConformanceFindings(modelRef, denied, allowedSet, policyDocName));
-  }
-
-  return findings;
-}
-
-function readModelProviderPolicyList(policy: unknown, path: readonly string[]): readonly string[] {
-  return readStringList(policy, path).map((provider) => normalizeProviderId(provider));
-}
-
-function modelProviderConformanceFindings(
-  provider: PolicyEvidence["modelProviders"][number],
-  denied: ReadonlySet<string>,
-  allowed: ReadonlySet<string>,
-  policyDocName: string,
-): readonly HealthFinding[] {
-  const findings: HealthFinding[] = [];
-  if (denied.has(provider.id)) {
-    findings.push({
-      checkId: CHECK_IDS.policyDeniedModelProvider,
-      severity: "error",
-      message: `Model provider '${provider.id}' is denied by policy.`,
-      source: "policy",
-      path: "openclaw config",
-      ocPath: provider.source,
-      target: provider.source,
-      requirement: `oc://${policyDocName}/models/providers/deny`,
-      fixHint: "Remove this configured provider or update the policy after review.",
-    });
-  }
-  if (!denied.has(provider.id) && allowed.size > 0 && !allowed.has(provider.id)) {
-    findings.push({
-      checkId: CHECK_IDS.policyUnapprovedModelProvider,
-      severity: "error",
-      message: `Model provider '${provider.id}' is not in the policy allowlist.`,
-      source: "policy",
-      path: "openclaw config",
-      ocPath: provider.source,
-      target: provider.source,
-      requirement: `oc://${policyDocName}/models/providers/allow`,
-      fixHint: "Use an approved model provider or update the policy after review.",
-    });
-  }
-  return findings;
-}
-
-function modelRefConformanceFindings(
-  modelRef: PolicyEvidence["modelRefs"][number],
-  denied: ReadonlySet<string>,
-  allowed: ReadonlySet<string>,
-  policyDocName: string,
-): readonly HealthFinding[] {
-  const findings: HealthFinding[] = [];
-  if (denied.has(modelRef.provider)) {
-    findings.push({
-      checkId: CHECK_IDS.policyDeniedModelProvider,
-      severity: "error",
-      message: `Model ref '${modelRef.ref}' uses denied provider '${modelRef.provider}'.`,
-      source: "policy",
-      path: "openclaw config",
-      ocPath: modelRef.source,
-      target: modelRef.source,
-      requirement: `oc://${policyDocName}/models/providers/deny`,
-      fixHint: "Select an approved model provider or update the policy after review.",
-    });
-  }
-  if (!denied.has(modelRef.provider) && allowed.size > 0 && !allowed.has(modelRef.provider)) {
-    findings.push({
-      checkId: CHECK_IDS.policyUnapprovedModelProvider,
-      severity: "error",
-      message: `Model ref '${modelRef.ref}' uses unapproved provider '${modelRef.provider}'.`,
-      source: "policy",
-      path: "openclaw config",
-      ocPath: modelRef.source,
-      target: modelRef.source,
-      requirement: `oc://${policyDocName}/models/providers/allow`,
-      fixHint: "Select an approved model provider or update the policy after review.",
-    });
-  }
-  return findings;
-}
-
-function networkFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  const allowPrivateNetwork = readPolicyBoolean(policy, ["network", "privateNetwork", "allow"]);
-  if (allowPrivateNetwork !== false) {
-    return [];
-  }
-  return evidence.network
-    .filter((setting) => setting.value)
-    .map((setting): HealthFinding => {
-      return {
-        checkId: CHECK_IDS.policyPrivateNetworkAccess,
-        severity: "error",
-        message: `Network setting '${setting.id}' allows private-network access.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: setting.source,
-        target: setting.source,
-        requirement: `oc://${policyDocName}/network/privateNetwork/allow`,
-        fixHint: "Disable this private-network access setting or update policy after review.",
-      };
-    });
-}
-
 function ingressFindings(
   policy: unknown,
   policyPath: string,
@@ -3968,239 +2527,6 @@ function ingressLabel(entry: PolicyIngressEvidence): string {
   const account = entry.accountId === undefined ? "" : ` account '${entry.accountId}'`;
   const group = entry.groupId === undefined ? "" : ` group '${entry.groupId}'`;
   return `channel '${entry.channel ?? "unknown"}'${account}${group}`;
-}
-
-function gatewayExposureFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  return [
-    ...gatewayNonLoopbackBindFindings(policy, policyDocName, evidence),
-    ...gatewayAuthFindings(policy, policyDocName, evidence),
-    ...gatewayControlUiFindings(policy, policyDocName, evidence),
-    ...gatewayTailscaleFindings(policy, policyDocName, evidence),
-    ...gatewayRemoteFindings(policy, policyDocName, evidence),
-    ...gatewayHttpEndpointFindings(policy, policyDocName, evidence),
-    ...gatewayHttpUrlFetchFindings(policy, policyDocName, evidence),
-  ];
-}
-
-function gatewayNonLoopbackBindFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  if (readPolicyBoolean(policy, ["gateway", "exposure", "allowNonLoopbackBind"]) !== false) {
-    return [];
-  }
-  return (evidence.gatewayExposure ?? [])
-    .filter((entry) => entry.kind === "bind" && entry.nonLoopback === true)
-    .map((entry): HealthFinding => {
-      return {
-        checkId: CHECK_IDS.policyGatewayNonLoopbackBind,
-        severity: "error",
-        message:
-          entry.explicit === false
-            ? "Gateway bind is omitted while the runtime default can permit non-loopback exposure."
-            : `Gateway bind setting '${entry.id}' permits non-loopback exposure.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: entry.source,
-        target: entry.source,
-        requirement: `oc://${policyDocName}/gateway/exposure/allowNonLoopbackBind`,
-        fixHint: "Use gateway.bind=loopback or update policy after review.",
-      };
-    });
-}
-
-function gatewayAuthFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  const findings: HealthFinding[] = [];
-  if (readPolicyBoolean(policy, ["gateway", "auth", "requireAuth"]) === true) {
-    findings.push(
-      ...(evidence.gatewayExposure ?? [])
-        .filter((entry) => entry.kind === "auth" && entry.value === "none")
-        .map((entry): HealthFinding => {
-          return {
-            checkId: CHECK_IDS.policyGatewayAuthDisabled,
-            severity: "error",
-            message: "Gateway authentication is disabled.",
-            source: "policy",
-            path: "openclaw config",
-            ocPath: entry.source,
-            target: entry.source,
-            requirement: `oc://${policyDocName}/gateway/auth/requireAuth`,
-            fixHint: "Set gateway.auth.mode to token, password, or trusted-proxy.",
-          };
-        }),
-    );
-  }
-  if (readPolicyBoolean(policy, ["gateway", "auth", "requireExplicitRateLimit"]) === true) {
-    findings.push(
-      ...(evidence.gatewayExposure ?? [])
-        .filter((entry) => entry.kind === "authRateLimit" && entry.explicit !== true)
-        .map((entry): HealthFinding => {
-          return {
-            checkId: CHECK_IDS.policyGatewayRateLimitMissing,
-            severity: "error",
-            message: "Gateway authentication rate-limit posture is not explicit.",
-            source: "policy",
-            path: "openclaw config",
-            ocPath: entry.source,
-            target: entry.source,
-            requirement: `oc://${policyDocName}/gateway/auth/requireExplicitRateLimit`,
-            fixHint: "Configure gateway.auth.rateLimit or update policy after review.",
-          };
-        }),
-    );
-  }
-  return findings;
-}
-
-function gatewayControlUiFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  if (readPolicyBoolean(policy, ["gateway", "controlUi", "allowInsecure"]) !== false) {
-    return [];
-  }
-  return (evidence.gatewayExposure ?? [])
-    .filter(
-      (entry) =>
-        entry.kind === "controlUi" &&
-        entry.value === true &&
-        (entry.id === "gateway-control-ui-insecure-auth" ||
-          entry.id === "gateway-control-ui-device-auth-disabled" ||
-          entry.id === "gateway-control-ui-host-origin-fallback"),
-    )
-    .map((entry): HealthFinding => {
-      return {
-        checkId: CHECK_IDS.policyGatewayControlUiInsecure,
-        severity: "error",
-        message: `Gateway Control UI insecure toggle '${entry.id}' is enabled.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: entry.source,
-        target: entry.source,
-        requirement: `oc://${policyDocName}/gateway/controlUi/allowInsecure`,
-        fixHint: "Disable the insecure Control UI toggle or update policy after review.",
-      };
-    });
-}
-
-function gatewayTailscaleFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  if (readPolicyBoolean(policy, ["gateway", "exposure", "allowTailscaleFunnel"]) !== false) {
-    return [];
-  }
-  return (evidence.gatewayExposure ?? [])
-    .filter((entry) => entry.kind === "tailscale" && entry.value === "funnel")
-    .map((entry): HealthFinding => {
-      return {
-        checkId: CHECK_IDS.policyGatewayTailscaleFunnel,
-        severity: "error",
-        message: "Gateway Tailscale Funnel exposure is enabled.",
-        source: "policy",
-        path: "openclaw config",
-        ocPath: entry.source,
-        target: entry.source,
-        requirement: `oc://${policyDocName}/gateway/exposure/allowTailscaleFunnel`,
-        fixHint: "Use tailscale serve/off or update policy after review.",
-      };
-    });
-}
-
-function gatewayRemoteFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  if (readPolicyBoolean(policy, ["gateway", "remote", "allow"]) !== false) {
-    return [];
-  }
-  return (evidence.gatewayExposure ?? [])
-    .filter((entry) => entry.kind === "remote")
-    .map((entry): HealthFinding => {
-      return {
-        checkId: CHECK_IDS.policyGatewayRemoteEnabled,
-        severity: "error",
-        message: `Gateway remote posture '${entry.id}' is enabled.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: entry.source,
-        target: entry.source,
-        requirement: `oc://${policyDocName}/gateway/remote/allow`,
-        fixHint: "Disable remote gateway mode/config or update policy after review.",
-      };
-    });
-}
-
-function gatewayHttpEndpointFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  const denied = new Set(
-    readStringList(policy, ["gateway", "http", "denyEndpoints"]).map((endpoint) =>
-      endpoint.toLowerCase(),
-    ),
-  );
-  if (denied.size === 0) {
-    return [];
-  }
-  return (evidence.gatewayExposure ?? [])
-    .filter(
-      (entry) =>
-        entry.kind === "httpEndpoint" &&
-        entry.endpoint !== undefined &&
-        denied.has(entry.endpoint.toLowerCase()),
-    )
-    .map((entry): HealthFinding => {
-      return {
-        checkId: CHECK_IDS.policyGatewayHttpEndpointEnabled,
-        severity: "error",
-        message: `Gateway HTTP endpoint '${entry.endpoint ?? entry.id}' is denied by policy.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: entry.source,
-        target: entry.source,
-        requirement: `oc://${policyDocName}/gateway/http/denyEndpoints`,
-        fixHint: "Disable the HTTP endpoint or update policy after review.",
-      };
-    });
-}
-
-function gatewayHttpUrlFetchFindings(
-  policy: unknown,
-  policyDocName: string,
-  evidence: PolicyEvidence,
-): readonly HealthFinding[] {
-  if (readPolicyBoolean(policy, ["gateway", "http", "requireUrlAllowlists"]) !== true) {
-    return [];
-  }
-  return (evidence.gatewayExposure ?? [])
-    .filter((entry) => entry.kind === "httpUrlFetch" && entry.hasAllowlist !== true)
-    .map((entry): HealthFinding => {
-      return {
-        checkId: CHECK_IDS.policyGatewayHttpUrlFetchUnrestricted,
-        severity: "error",
-        message: `Gateway HTTP URL-fetch input '${entry.id}' has no URL allowlist.`,
-        source: "policy",
-        path: "openclaw config",
-        ocPath: entry.source,
-        target: entry.source,
-        requirement: `oc://${policyDocName}/gateway/http/requireUrlAllowlists`,
-        fixHint: "Add a urlAllowlist for this URL-fetch input or update policy after review.",
-      };
-    });
 }
 
 function agentWorkspaceFindings(
@@ -6466,168 +4792,6 @@ function scopedPolicyFields(
     }));
 }
 
-export function isPolicyValueAtLeastAsStrict(
-  metadata: PolicyRuleMetadata,
-  candidate: unknown,
-  baseline: unknown,
-): boolean {
-  switch (metadata.strictness) {
-    case "allowlist-subset":
-      return isPolicyAllowlistSubset(metadata, candidate, baseline);
-    case "denylist-superset":
-      return isPolicyDenylistSuperset(metadata, candidate, baseline);
-    case "ordered-string":
-      return isPolicyOrderedStringAtLeastAsStrict(metadata, candidate, baseline);
-    case "requires-true":
-      return baseline !== true || candidate === true;
-    case "requires-false":
-      return baseline !== false || candidate === false;
-    case "exact-list":
-      return samePolicyStringList(candidate, baseline, metadata);
-  }
-  return false;
-}
-
-function isPolicyOrderedStringAtLeastAsStrict(
-  metadata: PolicyRuleMetadata,
-  candidate: unknown,
-  baseline: unknown,
-): boolean {
-  const candidateValue = policyString(candidate, metadata);
-  const baselineValue = policyString(baseline, metadata);
-  if (
-    candidateValue === undefined ||
-    baselineValue === undefined ||
-    metadata.orderedValues === undefined
-  ) {
-    return false;
-  }
-  const orderedValues = metadata.orderedValues.map((entry) =>
-    metadata.caseSensitive === true ? entry : entry.toLowerCase(),
-  );
-  const candidateIndex = orderedValues.indexOf(candidateValue);
-  const baselineIndex = orderedValues.indexOf(baselineValue);
-  return candidateIndex >= 0 && baselineIndex >= 0 && candidateIndex >= baselineIndex;
-}
-
-function isPolicyAllowlistSubset(
-  metadata: PolicyRuleMetadata,
-  candidate: unknown,
-  baseline: unknown,
-): boolean {
-  const candidateList = policyStringList(candidate, metadata);
-  const baselineList = policyStringList(baseline, metadata);
-  if (candidateList === undefined || baselineList === undefined) {
-    return false;
-  }
-  if (metadata.emptyList === "disabled" && baselineList.length === 0) {
-    return true;
-  }
-  if (metadata.emptyList === "disabled" && baselineList.length > 0 && candidateList.length === 0) {
-    return false;
-  }
-  const allowed = new Set(baselineList);
-  return candidateList.every((entry) => allowed.has(entry));
-}
-
-function isPolicyDenylistSuperset(
-  metadata: PolicyRuleMetadata,
-  candidate: unknown,
-  baseline: unknown,
-): boolean {
-  const candidateList = policyStringList(candidate, metadata);
-  const baselineList = policyStringList(baseline, metadata);
-  if (candidateList === undefined || baselineList === undefined) {
-    return false;
-  }
-  if (metadata.policyPath.join(".") === "tools.denyTools") {
-    return baselineList
-      .flatMap(expandPolicyToolRequirement)
-      .every((tool) => toolListCoversTool(candidateList, tool));
-  }
-  const denied = new Set(candidateList);
-  return baselineList.every((entry) => denied.has(entry));
-}
-
-function samePolicyStringList(
-  candidate: unknown,
-  baseline: unknown,
-  metadata: PolicyRuleMetadata,
-): boolean {
-  const candidateList = policyStringList(candidate, metadata);
-  const baselineList = policyStringList(baseline, metadata);
-  if (candidateList === undefined || baselineList === undefined) {
-    return false;
-  }
-  const candidateSorted = candidateList.toSorted();
-  const baselineSorted = baselineList.toSorted();
-  return (
-    candidateSorted.length === baselineSorted.length &&
-    candidateSorted.every((entry, index) => entry === baselineSorted[index])
-  );
-}
-
-function policyStringList(
-  value: unknown,
-  metadata: PolicyRuleMetadata,
-): readonly string[] | undefined {
-  if (metadata.valueType === "channel-provider-deny-rules") {
-    return channelProviderDenyRuleList(value, metadata);
-  }
-  if (!Array.isArray(value)) {
-    return undefined;
-  }
-  if (metadata.policyPath.join(".") === "execApprovals.agents.allowlist.expected") {
-    const entries = value.map(execApprovalAllowlistRequirement);
-    if (!entries.every((entry): entry is ExecApprovalAllowlistRequirement => entry !== undefined)) {
-      return undefined;
-    }
-    return entries.map((entry) => entry.key);
-  }
-  if (!value.every((entry) => typeof entry === "string")) {
-    return undefined;
-  }
-  return value
-    .map((entry) => entry.trim())
-    .filter(Boolean)
-    .map((entry) => normalizePolicyStringListEntry(entry, metadata));
-}
-
-function normalizePolicyStringListEntry(entry: string, metadata: PolicyRuleMetadata): string {
-  if (metadata.normalizeValues === "model-provider") {
-    return normalizeProviderId(entry);
-  }
-  return metadata.caseSensitive === true ? entry : entry.toLowerCase();
-}
-
-function channelProviderDenyRuleList(
-  value: unknown,
-  metadata: PolicyRuleMetadata,
-): readonly string[] | undefined {
-  if (!Array.isArray(value)) {
-    return undefined;
-  }
-  const providers: string[] = [];
-  for (const entry of value) {
-    if (!isChannelDenyRule(entry)) {
-      return undefined;
-    }
-    const provider = entry.when?.provider?.trim();
-    if (provider !== undefined && provider !== "") {
-      providers.push(metadata.caseSensitive === true ? provider : provider.toLowerCase());
-    }
-  }
-  return providers;
-}
-
-function policyString(value: unknown, metadata: PolicyRuleMetadata): string | undefined {
-  if (typeof value !== "string" || value.trim() === "") {
-    return undefined;
-  }
-  const trimmed = value.trim();
-  return metadata.caseSensitive === true ? trimmed : trimmed.toLowerCase();
-}
-
 function scopedPolicyValue(overlay: Record<string, unknown>, path: readonly string[]): unknown {
   const scopedRoot = path[0] === "agents" ? overlay.agents : overlay[path[0]];
   if (path[0] === "agents") {
@@ -7329,67 +5493,6 @@ function authProfileHasMetadata(
   return SUPPORTED_AUTH_PROFILE_MODES.includes(
     profile.mode as (typeof SUPPORTED_AUTH_PROFILE_MODES)[number],
   );
-}
-
-function readPolicyStringArray(
-  policy: unknown,
-  path: readonly string[],
-  options: { readonly lowercase?: boolean } = {},
-): readonly string[] | undefined {
-  let current: unknown = policy;
-  for (const part of path) {
-    if (!isRecord(current)) {
-      return undefined;
-    }
-    current = current[part];
-  }
-  if (!Array.isArray(current) || !current.every((entry) => typeof entry === "string")) {
-    return undefined;
-  }
-  const lowercase = options.lowercase ?? true;
-  return current
-    .map((entry) => {
-      const trimmed = entry.trim();
-      return lowercase ? trimmed.toLowerCase() : trimmed;
-    })
-    .filter(Boolean);
-}
-
-function readStringList(
-  policy: unknown,
-  path: readonly string[],
-  options?: { readonly lowercase?: boolean },
-): readonly string[] {
-  return readPolicyStringArray(policy, path, options) ?? [];
-}
-
-function readString(policy: unknown, path: readonly string[]): string | undefined {
-  let current: unknown = policy;
-  for (const part of path) {
-    if (!isRecord(current)) {
-      return undefined;
-    }
-    current = current[part];
-  }
-  return typeof current === "string" ? current.trim().toLowerCase() : undefined;
-}
-
-function ocPathSegment(value: string): string {
-  if (/^(?:[A-Za-z0-9_-]+|#\d+)$/.test(value)) {
-    return value;
-  }
-  return JSON.stringify(value);
-}
-
-function readPolicyBoolean(policy: unknown, path: readonly string[]): boolean | undefined {
-  let current: unknown = policy;
-  for (const part of path) {
-    if (!isRecord(current)) {
-      return undefined;
-    }
-    current = current[part];
-  }
-  return typeof current === "boolean" ? current : undefined;
 }
 
 function policyToolGlobMatches(tool: string, pattern: string): boolean {
